Better report of next cron run.
The JACS API returns the next run of a scheduled job, based on the last
time it was run, and it doesn't get updated if the cron is paused. This
code takes the cron string, parses it and decides when the next run wold
occur based on the current time.